pushbullet/requirements.txt
#requests requirement moved to core
python-magic>=0.4.12
